cannot use config.MetricsPort (type uint16) as type int in assignment

+++ app/controllers/metrics_controller.go
@@ -117,7 +117,7 @@ var (
 type MetricsController struct {
 	endpointsMap endpointsInfoMap
 	MetricsPath  string
-	MetricsPort  int
+	MetricsPort  uint16
 	mu           sync.Mutex
 	nodeIP       net.IP
 	serviceMap   serviceInfoMap
